Refactor NPC to a modular ESM architecture and improve UI/UX
- Modularized JavaScript into state, ui, generator, and constants modules.
- Implemented real-time prompt preview.
- Redesigned the long form into a collapsible accordion layout.
- Added character counters for all textareas.
- Implemented state persistence via localStorage.
- Cleaned up index.html and style.css for better maintainability and look.
