+++ style.css
@@ -58,6 +58,110 @@ h2 {
     padding-bottom: 0.75rem;
 }
 
+/* Accordion Styles */
+.accordion-item {
+    border: 1px solid #374151;
+    border-radius: 0.5rem;
+    margin-bottom: 1.25rem;
+    overflow: hidden;
+    background-color: #1f2937;
+    box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+}
+
+.accordion-header {
+    background-color: #1f2937;
+    color: #34d399;
+    padding: 1.25rem 1.5rem;
+    margin: 0 !important;
+    cursor: pointer;
+    font-size: 0.9rem !important;
+    border-bottom: none !important;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    transition: background-color 0.2s;
+    user-select: none;
+}
+
+.accordion-header:hover {
+    background-color: #374151;
+}
+
+.accordion-header::after {
+    content: '+';
+    font-family: monospace;
+    font-size: 1.25rem;
+    font-weight: bold;
+    transition: transform 0.3s;
+}
+
+.accordion-item.expanded .accordion-header::after {
+    content: '-';
+}
+
+.accordion-item.expanded .accordion-header {
+    border-bottom: 1px solid #374151 !important;
+    background-color: #374151;
+}
+
+.accordion-content {
+    max-height: 0;
+    overflow: hidden;
+    transition: max-height 0.3s ease-out;
+    background-color: #1f2937;
+}
+
+.accordion-item.expanded .accordion-content {
+    padding: 1.5rem;
+    max-height: 3000px;
+    transition: max-height 0.5s ease-in;
+}
+
+.help-text {
+    font-size: 0.8rem;
+    color: #9ca3af;
+    margin-top: 0.5rem;
+    font-style: italic;
+}
+
+.subgroup {
+    border-left: 4px solid #34d399;
+    padding: 1rem 1rem 1rem 1.5rem;
+    margin-top: 1.5rem;
+    background-color: #111827;
+    border-radius: 0 0.5rem 0.5rem 0;
+}
+
+.checkbox-group {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(220px, 1fr));
+    gap: 0.75rem;
+    background: #111827;
+    padding: 1rem;
+    border-radius: 0.375rem;
+    border: 1px solid #374151;
+}
+
+.checkbox-inline {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+}
+
+.checkbox-inline input[type="checkbox"] {
+    width: 1.25rem;
+    height: 1.25rem;
+    cursor: pointer;
+}
+
+.char-counter {
+    display: block;
+    text-align: right;
+    font-size: 0.75rem;
+    color: #9ca3af;
+    margin-top: 0.25rem;
+}
+
 .form-group {
     margin-bottom: 1.5rem;
 }
